-evaluatePremium

+++ src/main/java/org/camunda/bpm/entities/ContractEntity.java
@@ -64,7 +64,7 @@ public class ContractEntity implements Serializable {
 	public void setPrice(double price) {
 		this.price = price;
 	}
-	public double getDuration() {
+	public long getDuration() {
 		return duration;
 	}
 	public void setDuration(Long duration) {
